New business requirement: locator code
This code is made from ISBN, author name and number of pages.
It is an internal unique code used to keep control of the location of
a particular book.

To be able to create the code we'll use an external service to get the
data for a book given it's ISBN.
We'll cache the data in a database for reliability and performance. If
a code is not in the database, then te service will be queried.